Update PrepareTargetSubscriptionForCentralModeCA.md
Updated the script for preparing target subs in central scan CA mode. Now, the user needs to also supply the logging option as a parameter. And according to the logging option supplied, the script will set up the target sub's configurations. Also, fixed a bug regarding the check on sub access.